requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Stack
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
sentence-transformers==2.7.0
transformers==4.39.3
chromadb==0.4.22
openai==1.10.0
llama-index==0.10.12
llama-index-llms-openai
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-readers-file
hf_xet
pypdf==4.0.1

# Utilities
requests==2.31.0
httpx==0.26.0

# Frontend (Streamlit)
streamlit==1.30.0
plotly==5.18.0
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0

# Excel Support
openpyxl==3.1.2
xlrd==2.0.1
